+++ jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-api/src/test/java/jdplus/toolkit/base/api/timeseries/TsCollectionTest.java
@@ -1,23 +1,21 @@
 package jdplus.toolkit.base.api.timeseries;
 
 import _util.MockedTsProvider;
-import jdplus.toolkit.base.api.timeseries.*;
+import jdplus.toolkit.base.api.data.DoubleSeq;
 import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 import static jdplus.toolkit.base.api.timeseries.TsInformationType.BaseInformation;
 import static jdplus.toolkit.base.api.timeseries.TsInformationType.Data;
-import static java.util.Arrays.asList;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatNullPointerException;
+import static org.assertj.core.api.Assertions.*;
 
 public class TsCollectionTest {
 
+    @SuppressWarnings("DataFlowIssue")
     @Test
     public void testFactories() {
         assertThatNullPointerException()
@@ -37,7 +35,7 @@ public class TsCollectionTest {
         assertThat(Stream.<Ts>empty().collect(TsCollection.toTsCollection()))
                 .isEqualTo(TsCollection.EMPTY);
 
-        List<Ts> list = IntStream.range(0, 100).mapToObj(i -> Ts.builder().name("ts" + i).build()).collect(Collectors.toList());
+        List<Ts> list = IntStream.range(0, 100).mapToObj(i -> Ts.builder().name("ts" + i).build()).toList();
 
         assertThat(list.stream().collect(TsCollection.toTsCollection()))
                 .isEqualTo(TsCollection.of(list));
@@ -46,6 +44,7 @@ public class TsCollectionTest {
                 .isEqualTo(TsCollection.of(list));
     }
 
+    @SuppressWarnings("DataFlowIssue")
     @Test
     public void testLoadOfProvided() {
         TsCollection provided = factory.makeTsCollection(colMoniker, BaseInformation);
@@ -69,6 +68,7 @@ public class TsCollectionTest {
         }
     }
 
+    @SuppressWarnings("DataFlowIssue")
     @Test
     public void testLoadOfAnonymous() {
         TsCollection anonymous = TsCollection.of(Ts.of(TsData.empty("abc")));
@@ -95,25 +95,39 @@ public class TsCollectionTest {
         assertThat(col.toList()).containsExactlyElementsOf(col.getItems());
     }
 
+    @SuppressWarnings("DataFlowIssue")
     @Test
     public void testReplaceAll() {
-        assertThat(TsCollection.of(asList()).replaceAll(TsCollection.of(ts1b)))
+        assertThatNullPointerException()
+                .isThrownBy(() -> TsCollection.of(List.of()).replaceAll(null));
+
+        assertThat(TsCollection.of(List.of()).replaceAll(TsCollection.of(ts1b)))
                 .isEmpty();
 
-        assertThat(TsCollection.of(asList(ts1a)).replaceAll(TsCollection.of(ts1b)))
+        assertThat(TsCollection.of(List.of(ts1a)).replaceAll(TsCollection.of(ts1b)))
                 .containsExactly(ts1b);
 
-        assertThat(TsCollection.of(asList(ts1b)).replaceAll(TsCollection.of(ts1b)))
+        assertThat(TsCollection.of(List.of(ts1b)).replaceAll(TsCollection.of(ts1b)))
                 .containsExactly(ts1b);
 
-        assertThat(TsCollection.of(asList(ts2)).replaceAll(TsCollection.of(ts1b)))
+        assertThat(TsCollection.of(List.of(ts2)).replaceAll(TsCollection.of(ts1b)))
                 .containsExactly(ts2);
 
-        assertThat(TsCollection.of(asList(ts1a, ts2)).replaceAll(TsCollection.of(ts1b)))
+        assertThat(TsCollection.of(List.of(ts1a, ts2)).replaceAll(TsCollection.of(ts1b)))
                 .containsExactly(ts1b, ts2);
 
-        assertThat(TsCollection.of(asList(ts2, ts1a)).replaceAll(TsCollection.of(ts1b)))
+        assertThat(TsCollection.of(List.of(ts2, ts1a)).replaceAll(TsCollection.of(ts1b)))
                 .containsExactly(ts2, ts1b);
+    }
+
+    @SuppressWarnings("DataFlowIssue")
+    @Test
+    public void testOfName() {
+        assertThatNullPointerException()
+                .isThrownBy(() -> TsCollection.ofName(null));
+
+        assertThatObject(TsCollection.ofName(""))
+                .returns("", TsCollection::getName);
     }
 
     private final Ts ts1a = Ts.builder().name("ts1a").moniker(TsMoniker.of(MockedTsProvider.NAME, "0:1")).build();
@@ -128,6 +142,21 @@ public class TsCollectionTest {
                     .tsCollection(TsCollection.builder().moniker(colMoniker).build())
                     .build()
     );
+
+    @Test
+    public void testGetDomain() {
+        assertThat(TsCollection.EMPTY.getDomain())
+                .describedAs("No time series")
+                .isEqualTo(TsDomain.DEFAULT_EMPTY);
+
+        assertThat(TsCollection.of(Ts.of(TsData.of(TsPeriod.monthly(2010, 2), DoubleSeq.empty()))).getDomain())
+                .describedAs("Empty time series")
+                .isEqualTo(TsDomain.DEFAULT_EMPTY);
+
+        assertThat(TsCollection.of(Ts.of(TsData.of(TsPeriod.monthly(2010, 2), DoubleSeq.of(3.14)))).getDomain())
+                .describedAs("Non empty time series")
+                .hasToString("R1/2010-02-01T00:00:00/P1M");
+    }
 
     private final TsFactory factory = TsFactory.of(providers);
 }