+++ jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-api/src/test/java/jdplus/toolkit/base/api/timeseries/TsPeriodTest.java
@@ -204,30 +204,35 @@ public class TsPeriodTest {
         assertThat(TsPeriod.parse("2011-02-01T00:00/P1M"))
                 .hasToString("2011-02-01T00:00:00/P1M")
                 .returns("2011-02/P1M", HasShortStringRepresentation::toShortString)
+                .returns("2011-02", TsPeriod::getStartAsShortString)
                 .isEqualTo(TsPeriod.parse("2011-02/P1M"))
                 .isEqualTo(monthly(2011, 2));
 
         assertThat(TsPeriod.parse("2020-04-30T00:00/P7D"))
                 .hasToString("2020-04-30T00:00:00/P7D")
                 .returns("2020-04-30/P7D", HasShortStringRepresentation::toShortString)
+                .returns("2020-04-30", TsPeriod::getStartAsShortString)
                 .isEqualTo(TsPeriod.parse("2020-04-30/P7D"))
                 .isEqualTo(weekly(2020, 4, 30));
 
         assertThat(TsPeriod.parse("2011-04-01T00:00/P3M"))
                 .hasToString("2011-04-01T00:00:00/P3M")
                 .returns("2011-04/P3M", HasShortStringRepresentation::toShortString)
+                .returns("2011-04", TsPeriod::getStartAsShortString)
                 .isEqualTo(TsPeriod.parse("2011-04/P3M"))
                 .isEqualTo(quarterly(2011, 2));
 
         assertThat(TsPeriod.parse("2011-02-15T10:07/PT1M"))
                 .hasToString("2011-02-15T10:07:00/PT1M")
                 .returns("2011-02-15T10:07/PT1M", HasShortStringRepresentation::toShortString)
+                .returns("2011-02-15T10:07", TsPeriod::getStartAsShortString)
                 .isEqualTo(TsPeriod.parse("2011-02-15T10:07/PT1M"))
                 .isEqualTo(minutely(2011, 2, 15, 10, 7));
 
         assertThat(TsPeriod.parse("2011/P1M"))
                 .hasToString("2011-01-01T00:00:00/P1M")
                 .returns("2011-01/P1M", HasShortStringRepresentation::toShortString)
+                .returns("2011-01", TsPeriod::getStartAsShortString)
                 .isEqualTo(TsPeriod.parse("2011-01/P1M"))
                 .isEqualTo(monthly(2011, 1));