+++ jdplus-main-desktop/jdplus-toolkit-desktop-plugin/src/main/java/jdplus/toolkit/desktop/plugin/components/TsFeatureHelper.java
@@ -1,17 +1,17 @@
 /*
  * Copyright 2018 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package jdplus.toolkit.desktop.plugin.components;
@@ -20,15 +20,14 @@ import jdplus.toolkit.base.api.timeseries.Ts;
 import jdplus.toolkit.base.api.timeseries.TsData;
 import jdplus.toolkit.base.tsp.TsMeta;
 import lombok.AccessLevel;
-import org.checkerframework.checker.index.qual.NonNegative;
 import lombok.NonNull;
+import org.checkerframework.checker.index.qual.NonNegative;
 
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 
 /**
- *
  * @author Philippe Charles
  */
 @lombok.AllArgsConstructor(access = AccessLevel.PRIVATE)
@@ -60,20 +59,13 @@ public final class TsFeatureHelper {
     private final int[] endIndexes;
 
     public boolean hasFeature(@NonNull Feature feature, @NonNegative int series, @NonNegative int obs) {
-        switch (feature) {
-            case Backcasts:
-                return obs < begIndexes[series];
-            case Actual:
-                return begIndexes[series] <= obs && obs <= endIndexes[series];
-            case Forecasts:
-                return endIndexes[series] < obs;
-            case Public:
-                return false;
-            case Confidential:
-                return false;
-            default:
-                return false;
-        }
+        return switch (feature) {
+            case Backcasts -> obs < begIndexes[series];
+            case Actual -> begIndexes[series] <= obs && obs <= endIndexes[series];
+            case Forecasts -> endIndexes[series] < obs;
+            case Public -> false;
+            case Confidential -> false;
+        };
     }
 
     private static int getBegIndex(TsData data, LocalDateTime beg) {