+++ jdplus-main-desktop/jdplus-toolkit-desktop-plugin/src/main/java/jdplus/toolkit/desktop/plugin/core/components/ByAnnualFrequencyColumnGridData.java
@@ -22,6 +22,8 @@ import jdplus.toolkit.base.api.timeseries.TsData;
 import jdplus.toolkit.base.api.timeseries.TsDataTable;
 import jdplus.toolkit.base.api.timeseries.TsDomain;
 import ec.util.chart.ObsIndex;
+import lombok.NonNull;
+
 import java.time.Month;
 import java.time.format.TextStyle;
 
@@ -30,7 +32,7 @@ import java.util.Locale;
 /**
  * @author Philippe Charles
  */
-final class ByAnnualFrequencyColumn implements TsGridData {
+final class ByAnnualFrequencyColumnGridData implements TsGridData {
 
     private final int seriesIndex;
     private final TsData data;
@@ -40,7 +42,7 @@ final class ByAnnualFrequencyColumn implements TsGridData {
     private final int annualFrequency;
     private final TsGridObs obs;
 
-    public ByAnnualFrequencyColumn(Ts series, int seriesIndex) {
+    public ByAnnualFrequencyColumnGridData(Ts series, int seriesIndex) {
         this.seriesIndex = seriesIndex;
         this.data = series.getData();
         this.domain = series.getData().getDomain();
@@ -52,27 +54,27 @@ final class ByAnnualFrequencyColumn implements TsGridData {
 
     private int getPeriodId(int i, int j) {
         int periodId = j + (annualFrequency * i) - startPosition;
-        return (periodId < 0 || periodId >= domain.getLength()) ? -1 : periodId;
+        return (periodId < 0 || periodId >= domain.getLength()) ? NO_OBS_INDEX : periodId;
     }
 
     @Override
-    public String getRowName(int i) {
+    public @NonNull String getRowName(int i) {
         return Integer.toString(startYear + i);
     }
 
     @Override
-    public String getColumnName(int j) {
+    public @NonNull String getColumnName(int j) {
         return Month.of((j + 1) * 12 / annualFrequency).getDisplayName(TextStyle.SHORT_STANDALONE, Locale.getDefault(Locale.Category.DISPLAY));
     }
 
     @Override
-    public TsGridObs getObs(int i, int j) {
+    public @NonNull TsGridObs getObs(int i, int j) {
         int obsIndex = getPeriodId(i, j);
         obs.setIndex(obsIndex);
-        obs.setPeriod(obsIndex != -1 ? domain.get(obsIndex) : null);
+        obs.setPeriod(obsIndex != NO_OBS_INDEX ? domain.get(obsIndex) : null);
         obs.setSeriesIndex(seriesIndex);
-        obs.setStatus(obsIndex != -1 ? TsDataTable.ValueStatus.PRESENT : TsDataTable.ValueStatus.EMPTY);
-        obs.setValue(obsIndex != -1 ? data.getValue(obsIndex) : Double.NaN);
+        obs.setStatus(obsIndex != NO_OBS_INDEX ? TsDataTable.ValueStatus.PRESENT : TsDataTable.ValueStatus.EMPTY);
+        obs.setValue(obsIndex != NO_OBS_INDEX ? data.getValue(obsIndex) : Double.NaN);
         return obs;
     }
 
@@ -87,16 +89,16 @@ final class ByAnnualFrequencyColumn implements TsGridData {
     }
 
     @Override
-    public int getRowIndex(ObsIndex index) {
+    public int getRowIndex(@NonNull ObsIndex index) {
         return index.getSeries() != seriesIndex
-                ? -1
+                ? NO_OBS_INDEX
                 : (index.getObs() + startPosition) / annualFrequency;
     }
 
     @Override
-    public int getColumnIndex(ObsIndex index) {
+    public int getColumnIndex(@NonNull ObsIndex index) {
         return index.getSeries() != seriesIndex
-                ? -1
+                ? NO_OBS_INDEX
                 : (index.getObs() + startPosition) % annualFrequency;
     }
 }