+++ CHANGELOG.md
@@ -12,6 +12,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![STAT] Add support of compact notation in ISO 8601 time interval [#159](https://github.com/jdemetra/jdplus-main/issues/159)
 - ![OTHER] Add Windows aarch_64 to standalone binaries [#659](https://github.com/jdemetra/jdplus-main/issues/659)
 
+### Changed
+
+- ![UI] Unify period display in GUI [#489](https://github.com/jdemetra/jdplus-main/issues/489)
+
 ## [3.5.1] - 2025-06-12
 
 ### Fixed